# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/vendors": {
      "get": {
        "operationId": "VendorsController_listVendors",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "This organization's vendors and Aveline's curated list, optionally by category",
        "tags": [
          "vendors"
        ]
      },
      "post": {
        "description": "To the caller's organization, which alone sees it; Aveline staff add to the curated list.",
        "operationId": "VendorsController_createVendor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVendorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Add a vendor",
        "tags": [
          "vendors"
        ]
      }
    },
    "/api/v1/events/{eventId}/vendors": {
      "get": {
        "description": "Fees only for callers holding vendor:fee:read; each brief link only for callers holding vendor:write.",
        "operationId": "VendorsController_listBookings",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "This event's vendors",
        "tags": [
          "vendors"
        ]
      },
      "post": {
        "description": "Scopes default to the category’s usual set, so forgetting them narrows what the vendor sees rather than widening it.",
        "operationId": "VendorsController_book",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookVendorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Engage a vendor and mint their brief link",
        "tags": [
          "vendors"
        ]
      }
    },
    "/api/v1/events/{eventId}/vendors/{bookingId}": {
      "patch": {
        "operationId": "VendorsController_updateBooking",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bookingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Change a booking’s status, scopes or fee",
        "tags": [
          "vendors"
        ]
      },
      "delete": {
        "description": "Kept as a record, with the brief link rotated so it stops working.",
        "operationId": "VendorsController_cancelBooking",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bookingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Cancel an engagement",
        "tags": [
          "vendors"
        ]
      }
    },
    "/api/v1/events/{eventId}/vendors/{bookingId}/rotate-brief": {
      "post": {
        "description": "The only revocation a capability URL has. The old link stops working.",
        "operationId": "VendorsController_rotate",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bookingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Replace the brief link",
        "tags": [
          "vendors"
        ]
      }
    },
    "/api/v1/briefs/{briefToken}": {
      "get": {
        "description": "Contains only the sections the booking lists, and says which those are.",
        "operationId": "VendorsController_findBrief",
        "parameters": [
          {
            "name": "briefToken",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "A vendor’s brief",
        "tags": [
          "vendors"
        ]
      }
    },
...
  },
  "info": {
    "title": "Aveline API",
    "description": "Event invitations, guest graph, ticketing and operations. See docs/API.md for conventions that apply across every endpoint.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateVendorDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "VENUE",
              "CATERING",
              "BAR",
              "DECOR",
              "PHOTOGRAPHY",
              "VIDEOGRAPHY",
              "MUSIC",
              "PRINT",
              "OTHER"
            ]
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "category"
        ]
      },
      "BookVendorDto": {
        "type": "object",
        "properties": {
          "vendorId": {
            "type": "string"
          },
          "briefScopes": {
            "type": "array",
            "description": "Exactly what this vendor may read. Omitted uses the category default.",
            "items": {
              "type": "string",
              "enum": [
                "headcount",
                "catering",
                "bar",
                "playlist",
                "timeline",
                "seating",
                "households",
                "contacts"
              ]
            }
          },
          "feeAmount": {
            "type": "string",
            "description": "Agreed fee, as a decimal string",
            "example": "150000.00"
          }
        },
        "required": [
          "vendorId"
        ]
      },
      "UpdateBookingDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ENQUIRED",
              "QUOTED",
              "CONFIRMED",
              "CANCELLED",
              "COMPLETED"
            ]
          },
          "briefScopes": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "headcount",
                "catering",
                "bar",
                "playlist",
                "timeline",
                "seating",
                "households",
                "contacts"
              ]
            }
          },
          "feeAmount": {
            "type": "string"
          }
        }
      },
...
    }
  }
}